+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [22.6.0] - Aspose Words Cloud for Dart 22.6 Release Notes
+
+- Added 'DeleteBookmark' and 'DeleteBookmarkOnline' API methods for delete bookmarks by name from the document.
+- Added 'DeleteBookmarks' and 'DeleteBookmarksOnline' API methods for delete all bookmarks from the document.
+- Added 'InsertBookmark' and 'InsertBookmarkOnline' API methods for create new bookmarks in the document.
+
+
 ## [22.5.0] - Aspose Words Cloud for Dart 22.5 Release Notes
 
 - Internal API fixes and improvments.